requirements.txt
# Google APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
google-generativeai>=0.8.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async
aiohttp>=3.9.0
asyncio>=3.4.3

# Utils
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.3
cachetools>=5.3.2
